Show comment count on tasks tile on landing page
