+++ packages/datatrak-web/src/features/Tasks/CommentsCount.tsx
@@ -1,0 +1,34 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+
+import React from 'react';
+import styled from 'styled-components';
+import { Typography } from '@material-ui/core';
+import { CommentIcon } from '../../components';
+
+const CommentsCountWrapper = styled.div`
+  color: ${({ theme }) => theme.palette.text.secondary};
+  display: flex;
+  align-items: center;
+  right: 0;
+  .MuiSvgIcon-root {
+    font-size: 1rem;
+  }
+`;
+
+const CommentCountText = styled(Typography)`
+  font-size: 0.75rem;
+  margin-inline-start: 0.25rem;
+`;
+
+export const CommentsCount = ({ commentsCount }: { commentsCount: number }) => {
+  if (!commentsCount) return null;
+  return (
+    <CommentsCountWrapper title="Number of user generated comments on this task">
+      <CommentIcon />
+      <CommentCountText>{commentsCount}</CommentCountText>
+    </CommentsCountWrapper>
+  );
+};
