+++ packages/datatrak-web/src/features/Tasks/TaskTile.tsx
@@ -11,6 +11,7 @@ import { ROUTES } from '../../constants';
 import { StatusPill } from './StatusPill';
 import { displayDate } from '../../utils';
 import { ButtonLink } from '../../components';
+import { CommentsCount } from './CommentsCount';
 
 const TileContainer = styled.div`
   display: flex;
@@ -63,28 +64,6 @@ const TileRight = styled.div`
   justify-content: center;
 `;
 
-const CommentsContainer = styled.div`
-  display: flex;
-  align-items: center;
-
-  .MuiSvgIcon-root {
-    font-size: 1rem;
-    margin-inline-end: 0.2rem;
-  }
-`;
-
-const Comments = ({ commentsCount = 0 }) => {
-  if (!commentsCount) {
-    return null;
-  }
-  return (
-    <CommentsContainer>
-      <ChatIcon />
-      {commentsCount}
-    </CommentsContainer>
-  );
-};
-
 export const TaskTile = ({ task }) => {
   const { survey, entity, taskStatus, dueDate } = task;
   const surveyLink = generatePath(ROUTES.SURVEY, {
@@ -98,7 +77,7 @@ export const TaskTile = ({ task }) => {
         <TileContent>
           <StatusPill status={taskStatus} />
           <span>{displayDate(dueDate)}</span>
-          <Comments />
+          <CommentsCount commentsCount={task.commentsCount} />
         </TileContent>
       </TileLeft>
       <TileRight>
